Navbar moved to home page
