+++ src/components/Hero.jsx
@@ -3,7 +3,7 @@ import {Link} from 'react-router-dom';
 
 const Hero = () => {
     return (
-        <div className=' hero-image bkground h-screen relative bg-black -top-28 -z-10'>
+        <div className=' hero-image bkground h-screen relative bg-black -top-28'>
             <section className='h-screen items-center relative md:flex md:flex-col md:justify-center'>
                 <div className='absolute bottom-16 pl-12 flex flex-col'>
                     <h1 className='hero-text font-black mt-10 text-white md:text-center md:text-6xl lg:text-center lg:text-7xl text-left text-4xl'>
