+++ src/pages/Home.jsx
@@ -1,11 +1,15 @@
+import Nav from '../components/homeNav';
 import Hero from '../components/Hero';
 import SectionOne from '../components/homeSectionOne';
 import SectionTwo from '../components/homeSectionTwo';
 import SectionThree from '../components/homeSectionThree';
 
 const Home = () => {
+
+
     return (
         <div>
+            <Nav />
             <Hero />
             <SectionOne />
             <SectionTwo />
